+++ src/ultimate_notion/obj_api/objects.py
@@ -490,6 +490,7 @@ class MentionLink(MentionBase, type='link_mention'):
 class BotTypeData(GenericObject):
     """Type data for a `Bot`."""
 
+    workspace_id: UUID
     owner: WorkspaceRef | MentionUser | None = None
     workspace_name: str | None = None
     workspace_limits: WorkSpaceLimits = Field(default_factory=WorkSpaceLimits)
@@ -501,7 +502,7 @@ class Bot(User[BotTypeData], type='bot'):
     # Even if stated otherwise in the docs, `bot` type data is optional and for instance
     # not present when a new page is created by a bot within a database with a `CreatedBy` Property.
     # For ease of use, we include a default instance of the bot type data.
-    bot: BotTypeData = Field(default_factory=BotTypeData)
+    bot: BotTypeData = Field(default_factory=BotTypeData)  # type: ignore[arg-type]
 
 
 class MentionPage(MentionBase[ObjectRef], type='page'):